Refactor into packageutils with retry
