+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "sfpowerkit",
   "description": "A Salesforce DX Plugin with multiple functionalities aimed at improving development and operational workflows",
-  "version": "2.0.10",
+  "version": "2.0.12",
   "author": {
     "name": "DXatScale",
     "email": "dxatscale@accenture.com"
