+++ src/impl/package/version/packageInfo.ts
@@ -1,6 +1,10 @@
 import { SFPowerkit } from "../../../sfpowerkit";
 import { Connection } from "jsforce";
 import { LoggerLevel } from "@salesforce/core";
+import {
+  getInstalledPackages,
+  PackageDetail
+} from "../../../utils/packageUtils";
 
 export default class PackageInfo {
   conn: Connection;
@@ -17,90 +21,12 @@ export default class PackageInfo {
 
   public async getPackages(): Promise<PackageDetail[]> {
     //await this.getInstalledPackageInfo();
-    let packageDetails = await this.getInstalledPackages();
+    let packageDetails = await getInstalledPackages(this.conn, true);
+
     SFPowerkit.log(
       "PackageDetails:" + JSON.stringify(packageDetails),
       LoggerLevel.TRACE
     );
     return packageDetails;
   }
-
-  private async getInstalledPackages(): Promise<PackageDetail[]> {
-    let packageDetails = [];
-    let installedPackagesQuery =
-      "SELECT Id, SubscriberPackageId, SubscriberPackage.NamespacePrefix, SubscriberPackage.Name, " +
-      "SubscriberPackageVersion.Id, SubscriberPackageVersion.Name, SubscriberPackageVersion.MajorVersion, SubscriberPackageVersion.MinorVersion, " +
-      "SubscriberPackageVersion.PatchVersion, SubscriberPackageVersion.BuildNumber FROM InstalledSubscriberPackage " +
-      "ORDER BY SubscriberPackageId";
-
-    let packageNamespacePrefixList = [];
-    await this.conn.tooling.query(installedPackagesQuery).then(queryResult => {
-      const records = queryResult.records;
-
-      if (records && records.length > 0) {
-        records.forEach(record => {
-          const packageDetail = {} as PackageDetail;
-          packageDetail.packageName = record["SubscriberPackage"]["Name"];
-          packageDetail.subcriberPackageId = record["SubscriberPackageId"];
-          packageDetail.packageNamespacePrefix =
-            record["SubscriberPackage"]["NamespacePrefix"];
-          packageDetail.packageVersionId =
-            record["SubscriberPackageVersion"]["Id"];
-          packageDetail.packageVersionNumber = `${record["SubscriberPackageVersion"]["MajorVersion"]}.${record["SubscriberPackageVersion"]["MinorVersion"]}.${record["SubscriberPackageVersion"]["PatchVersion"]}.${record["SubscriberPackageVersion"]["BuildNumber"]}`;
-          packageDetails.push(packageDetail);
-          if (packageDetail.packageNamespacePrefix) {
-            packageNamespacePrefixList.push(
-              "'" + packageDetail.packageNamespacePrefix + "'"
-            );
-          }
-        });
-      }
-    });
-
-    let licenseMap = new Map();
-    if (packageNamespacePrefixList.length > 0) {
-      let packageLicensingQuery = `SELECT AllowedLicenses, UsedLicenses,ExpirationDate, NamespacePrefix, IsProvisioned, Status FROM PackageLicense  WHERE NamespacePrefix IN (${packageNamespacePrefixList})`;
-      await this.conn.query(packageLicensingQuery).then(queryResult => {
-        if (queryResult.records && queryResult.records.length > 0) {
-          queryResult.records.forEach(record => {
-            let licenseDetailObj = {} as PackageDetail;
-            licenseDetailObj.allowedLicenses = record["AllowedLicenses"];
-            licenseDetailObj.usedLicenses = record["UsedLicenses"];
-            licenseDetailObj.expirationDate = record["ExpirationDate"];
-            licenseDetailObj.status = record["Status"];
-            licenseMap.set(record["NamespacePrefix"], licenseDetailObj);
-          });
-        }
-      });
-    }
-
-    if (packageDetails && licenseMap) {
-      packageDetails.forEach(detail => {
-        if (
-          detail.packageNamespacePrefix &&
-          licenseMap.has(detail.packageNamespacePrefix)
-        ) {
-          let licDetail = licenseMap.get(detail.packageNamespacePrefix);
-          detail.allowedLicenses = licDetail.allowedLicenses;
-          detail.usedLicenses = licDetail.usedLicenses;
-          detail.expirationDate = licDetail.expirationDate;
-          detail.status = licDetail.status;
-        }
-      });
-    }
-
-    return packageDetails;
-  }
-}
-
-export interface PackageDetail {
-  packageName: string;
-  subcriberPackageId: string;
-  packageNamespacePrefix: string;
-  packageVersionNumber: string;
-  packageVersionId: string;
-  allowedLicenses: number;
-  usedLicenses: number;
-  expirationDate: string;
-  status: string;
 }
